Restructure class inheritance
